Split TextWrapper into its own file.
