+++ lib/chef_apply/error.rb
@@ -67,4 +67,3 @@ module ChefApply
   class APIError < Error
   end
 end
-
